Remove Stones to Minimize the Total
This Java program minimizes the total number of stones in a pile by repeatedly removing half of the largest pile's stones k times. It uses a Max Heap (PriorityQueue in descending order) to always pick the largest pile. In each iteration, the largest element is removed, reduced by half (using Math.floor(val/2)), and reinserted into the heap. This process continues for k operations. After all operations, the remaining stones are summed up and returned. The use of a Max Heap ensures efficient retrieval and modification of the largest element, making the time complexity O(k log N) due to heap operations. However, there is a bug where the heap can grow indefinitely due to always reinserting values, which needs to be fixed.